requirements.txt
python-telegram-bot[job-queue]>=20.0
apscheduler==3.10.4
pytz==2024.1
hijri-converter
sqlalchemy[asyncio]
asyncpg
redis
celery[redis]
yt-dlp
python-dotenv
openai-whisper
ffmpeg-python
Pillow
mutagen
hachoir
psycopg2-binary
requests
fastapi
uvicorn
PyJWT
openai
python-multipart
